+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2026-02-16
+
+### Added
+
+- "Liberating" status (in progress)
+
+### Changed
+
+- Major UI and styling overhaul
+
 ## [0.1.2] - 2026-02-16
 
 ### Changed
@@ -32,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Buttons for clearing liberation and LocalStorage
 - "LIBERATED" stamp for completed Warbonds
 
-[unreleased]: https://github.com/johnarp/warbond-tracker/compare/v0.1.1...HEAD
-[0.1.1]: https://github.com/johnarp/warbond-tracker/compare/v0.1.0...v0.1.1
-[0.1.0]: https://github.com/johnarp/warbond-tracker/releases/tag/v0.1.0
+[0.2.0]: https://github.com/johnarp/warbond-tracker/compare/0.1.2...v0.2.0
+[0.1.2]: https://github.com/johnarp/warbond-tracker/compare/0.1.1...0.1.2
+[0.1.1]: https://github.com/johnarp/warbond-tracker/compare/0.1.0...0.1.1
+[0.1.0]: https://github.com/johnarp/warbond-tracker/releases/tag/0.1.0